+++ CheckedExceptions.CodeFixes/AddCatchClauseToTryCodeFixProvider.cs
@@ -20,7 +20,7 @@ public class AddCatchClauseToTryCodeFixProvider : CodeFixProvider
         [CheckedExceptionsAnalyzer.DiagnosticIdUnhandled];
 
     public sealed override FixAllProvider GetFixAllProvider() =>
-        null!;
+        WellKnownFixAllProviders.BatchFixer;
 
     public sealed override async Task RegisterCodeFixesAsync(CodeFixContext context)
     {